Add permalink during compilation instead of rendering
The #11 PR made clear there was something wrong generating tokens in
the rendering function (since markdown-it gives the token constructor
only in the rule but not the renderer, and this totally makes sense).

Now 'renderPermalink' is called at compile time, giving a chance for
other plugins to transform the link itself, removing the 'opts.Token'
workaround, and the 'getAttr' function.